feat: finishing admin service detail
